drivers/ukintctlr/gic: Adapt following ACPI code move to drivers
Straightforward update according to migration of ACPI code to
drivers/firmware.
